backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-jose[cryptography]==3.3.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.3
httpx==0.19.0
python-multipart==0.0.5
pytz==2021.3
PyJWT==2.8.0
APScheduler==3.10.1
